Add files via upload
NITEMARE is the font used across the screens.
The QuitButton script is a misnomer. It is responsible for the functionality of every button across every scene.
spr_Key is the key sprite used for buttons.
WinLoss is a script used to transition the game to the GameComplete/GameOver scene depending on whether the win or loss condition(determined by global booleans gameWin and gameLose respectively) was met.